+++ app/src/main/java/org/lineageos/glimpse/fragments/ReelsFragment.kt
@@ -5,8 +5,10 @@
 
 package org.lineageos.glimpse.fragments
 
+import android.content.Intent
 import android.content.res.Configuration
 import android.os.Bundle
+import android.provider.MediaStore
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.os.bundleOf
@@ -18,13 +20,13 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.appbar.AppBarLayout
 import com.google.android.material.shape.MaterialShapeDrawable
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import org.lineageos.glimpse.R
+import org.lineageos.glimpse.ViewActivity
 import org.lineageos.glimpse.ext.getViewProperty
 import org.lineageos.glimpse.recyclerview.ThumbnailAdapter
 import org.lineageos.glimpse.recyclerview.ThumbnailLayoutManager
@@ -50,11 +52,6 @@ class ReelsFragment : Fragment(R.layout.fragment_reels) {
     private val appBarLayout by getViewProperty<AppBarLayout>(R.id.appBarLayout)
     private val reelsRecyclerView by getViewProperty<RecyclerView>(R.id.reelsRecyclerView)
 
-    // Fragments
-    private val parentNavController by lazy {
-        requireParentFragment().requireParentFragment().findNavController()
-    }
-
     // Permissions
     private val permissionsUtils by lazy { PermissionsUtils(requireContext()) }
     private val permissionsGatedCallback = PermissionsGatedCallback(this) {
@@ -74,9 +71,14 @@ class ReelsFragment : Fragment(R.layout.fragment_reels) {
     // MediaStore
     private val thumbnailAdapter by lazy {
         ThumbnailAdapter { media ->
-            parentNavController.navigate(
-                R.id.action_mainFragment_to_mediaViewerFragment,
-                MediaViewerFragment.createBundle(media, MediaStoreBuckets.MEDIA_STORE_BUCKET_REELS.id)
+            startActivity(
+                Intent(requireContext(), ViewActivity::class.java).apply {
+                    action = MediaStore.ACTION_REVIEW
+                    data = media.externalContentUri
+                    putExtra(
+                        ViewActivity.KEY_ALBUM_ID, MediaStoreBuckets.MEDIA_STORE_BUCKET_REELS.id
+                    )
+                }
             )
         }
     }
