+++ app/src/main/java/org/lineageos/glimpse/utils/PermissionsGatedCallback.kt
@@ -5,27 +5,54 @@
 
 package org.lineageos.glimpse.utils
 
+import android.app.Activity
+import android.content.Context
 import android.widget.Toast
+import androidx.activity.result.ActivityResultCaller
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import org.lineageos.glimpse.R
 
 /**
  * A class that checks main app permissions before starting the callback.
  */
-class PermissionsGatedCallback(
-    fragment: Fragment, private val callback: () -> Unit
+class PermissionsGatedCallback private constructor(
+    caller: ActivityResultCaller,
+    private val getContext: () -> Context,
+    private val getActivity: () -> Activity,
+    private val callback: () -> Unit,
 ) {
-    private val permissionsUtils by lazy { PermissionsUtils(fragment.requireContext()) }
-    private val mainPermissionsRequestLauncher = fragment.registerForActivityResult(
+    constructor(
+        fragment: Fragment,
+        callback: () -> Unit,
+    ) : this(
+        fragment,
+        { fragment.requireContext() },
+        { fragment.requireActivity() },
+        callback,
+    )
+
+    constructor(
+        activity: AppCompatActivity,
+        callback: () -> Unit,
+    ) : this(
+        activity,
+        { activity },
+        { activity },
+        callback,
+    )
+
+    private val permissionsUtils by lazy { PermissionsUtils(getContext()) }
+    private val mainPermissionsRequestLauncher = caller.registerForActivityResult(
         ActivityResultContracts.RequestMultiplePermissions()
     ) {
         if (it.isNotEmpty()) {
             if (!permissionsUtils.mainPermissionsGranted()) {
                 Toast.makeText(
-                    fragment.requireContext(), R.string.app_permissions_toast, Toast.LENGTH_SHORT
+                    getContext(), R.string.app_permissions_toast, Toast.LENGTH_SHORT
                 ).show()
-                fragment.requireActivity().finish()
+                getActivity().finish()
             } else {
                 callback()
             }
