--- app/src/main/res/layout/fragment_media_viewer.xml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-     SPDX-FileCopyrightText: 2023 The LineageOS Project
-     SPDX-License-Identifier: Apache-2.0
--->
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@android:color/black"
-    android:theme="@style/Theme.Glimpse.MediaViewer"
-    tools:context=".fragments.MediaViewerFragment">
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/viewPager"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/topSheetConstraintLayout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:background="@drawable/bg_media_viewer_top_sheet"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageButton
-            android:id="@+id/backButton"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:layout_marginTop="20dp"
-            android:layout_marginBottom="20dp"
-            android:layout_marginStart="16dp"
-            android:backgroundTint="@android:color/transparent"
-            android:src="@drawable/ic_back"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:tint="?attr/colorOnSurface" />
-
-        <TextView
-            android:id="@+id/dateTextView"
-            style="@style/Theme.Glimpse.MediaViewer.DateTimeText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="24dp"
-            android:layout_marginTop="16dp"
-            android:textAllCaps="true"
-            android:textColor="?attr/colorOnSurface"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/timeTextView"
-            style="@style/Theme.Glimpse.MediaViewer.DateTimeText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            android:layout_marginEnd="24dp"
-            android:textAllCaps="true"
-            android:textColor="?attr/colorOnSurface"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/dateTextView" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <LinearLayout
-        android:id="@+id/bottomSheetLinearLayout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:background="@drawable/bg_media_viewer_bottom_sheet"
-        android:gravity="center"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <ImageButton
-            android:id="@+id/favoriteButton"
-            style="@style/Theme.Glimpse.MediaViewer.BottomSheet.Button"
-            android:src="@drawable/ic_star_toggle" />
-
-        <ImageButton
-            android:id="@+id/shareButton"
-            style="@style/Theme.Glimpse.MediaViewer.BottomSheet.Button"
-            android:src="@drawable/ic_share" />
-
-        <ImageButton
-            android:id="@+id/infoButton"
-            style="@style/Theme.Glimpse.MediaViewer.BottomSheet.Button"
-            android:src="@drawable/ic_info" />
-
-        <ImageButton
-            android:id="@+id/adjustButton"
-            style="@style/Theme.Glimpse.MediaViewer.BottomSheet.Button"
-            android:src="@drawable/ic_tune" />
-
-        <ImageButton
-            android:id="@+id/deleteButton"
-            style="@style/Theme.Glimpse.MediaViewer.BottomSheet.Button"
-            android:src="@drawable/ic_delete" />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
