+++ app/src/main/res/navigation/app_navigation.xml
@@ -15,14 +15,6 @@
         tools:layout="@layout/fragment_main">
 
         <action
-            android:id="@+id/action_mainFragment_to_mediaViewerFragment"
-            app:destination="@id/mediaViewerFragment"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
-
-        <action
             android:id="@+id/action_mainFragment_to_albumFragment"
             app:destination="@id/albumFragment"
             app:enterAnim="@anim/nav_default_enter_anim"
@@ -32,21 +24,7 @@
     </fragment>
 
     <fragment
-        android:id="@+id/mediaViewerFragment"
-        android:name="org.lineageos.glimpse.fragments.MediaViewerFragment"
-        tools:layout="@layout/fragment_media_viewer" />
-
-    <fragment
         android:id="@+id/albumFragment"
         android:name="org.lineageos.glimpse.fragments.AlbumFragment"
-        tools:layout="@layout/fragment_album">
-
-        <action
-            android:id="@+id/action_albumFragment_to_mediaViewerFragment"
-            app:destination="@id/mediaViewerFragment"
-            app:enterAnim="@anim/nav_default_enter_anim"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
-    </fragment>
+        tools:layout="@layout/fragment_album" />
 </navigation>
